Removed some spurious float casts (explanation: float is implicitly promoted to double anyway, it's only useful for storage; also a fp variable forces conversion of other int in the expression without needing a cast) Typo fixes in some comments 'floatting' -> 'floating' :D

+++ pcbnew/dimension.cpp
@@ -325,7 +325,7 @@ static void BuildDimension( EDA_DRAW_PANEL* aPanel, wxDC* aDC,
     {
         wxPoint delta;
         int dx, dy;
-        float angle, depl;
+        double angle, depl;
         delta = Dimension->m_featureLineDO - Dimension->m_featureLineGO;
 
         /* Calculating the direction of travel perpendicular to the selected axis. */
